refactor(pkg): move CompareFloat func into pkg package
